Modification de la methode produire dans toutes les unités de productions Ajout de la methode getQuantite dans Vendeur Acheteur Modification getLImite producteur

+++ src/main/java/marche/traitement/Producteurs/Apiculteur.java
@@ -8,7 +8,7 @@ public class Apiculteur extends Producteur  {
     /**
      * Default constructor
      */
-    public Apiculteur(double solde,String nom,int limite)
+    public Apiculteur(double solde,String nom,double limite)
     {
         super(solde,nom,limite);
