+++ src/main/java/marche/traitement/Producteurs/Arboriculteur.java
@@ -8,7 +8,7 @@ public class Arboriculteur extends Producteur {
     /**
      * Default constructor
      */
-    public Arboriculteur(double solde,String nom,int limite) {
+    public Arboriculteur(double solde,String nom,double limite) {
         super(solde,nom,limite);
     }
