+++ src/main/java/marche/traitement/Producteurs/ProducteurDeViande.java
@@ -8,7 +8,7 @@ public class ProducteurDeViande extends Producteur {
     /**
      * Default constructor
      */
-    public ProducteurDeViande(double solde,String nom,int limite) {
+    public ProducteurDeViande(double solde,String nom,double limite) {
         super(solde,nom, limite);
     }
