+++ src/main/java/marche/traitement/Producteurs/ProducteurLaitier.java
@@ -8,7 +8,7 @@ public class ProducteurLaitier extends Producteur {
     /**
      * Default constructor
      */
-    public ProducteurLaitier(double solde,String nom, int limite) {
+    public ProducteurLaitier(double solde,String nom, double limite) {
         super(solde,nom,limite);
     }
